requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.1.3
crewai==0.80.0
crewai-tools==0.17.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
pillow==11.0.0
pytesseract==0.3.13
serpapi==0.1.5
google-search-results==2.4.2
gunicorn==23.0.0
